继续完善虚拟纹理的逻辑

+++ src/9.other/9.10.VirtualTexture/TileGenerator.cpp
@@ -58,7 +58,7 @@ bool TileGenerator::generate(const std::string& _filePath)
 
 	// Generate cache filename
 	char tmp[256];
-	std::snprintf(tmp, sizeof(tmp), "%.*s.vt", baseName.length(), baseName.c_str());
+	std::snprintf(tmp, sizeof(tmp), "%.*s.vt", (unsigned int)baseName.length(), baseName.c_str());
 
 	std::string cacheFilePath = "temp/";
 	cacheFilePath = cacheFilePath + tmp;
